# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SureWaka API
  description: |
    SureWaka logistics platform API. Connects senders with verified logistics providers
    and independent drivers across Nigeria using a multi-leg delivery model.

    **Auth:** All endpoints require a Clerk JWT Bearer token unless marked `security: []`.

    **Response envelope:** Every response is `{ data, error, meta }` where exactly one of
    data/error is non-null. meta carries pagination when applicable.

    **Currency:** All monetary values are in kobo (1 NGN = 100 kobo).
  version: "1.0.0"
servers:
  - url: https://api.surewaka.com/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Development
security:
  - BearerAuth: []
...
paths:
...
  /driver/location:
    post:
      tags: [Driver Location]
      summary: Update driver GPS position (real-time matching)
      description: |
        Stores the driver's position in Redis (geo set + metadata hash) for real-time
        spatial queries, publishes to Ably for live tracking, and optionally persists
        to Postgres when an active delivery is in progress.

        Rate-limited to 1 request per 2 seconds per driver.
        Requires `driver` or `carrier_driver` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat: { type: number, minimum: -90, maximum: 90 }
                lng: { type: number, minimum: -180, maximum: 180 }
                deliveryId:
                  type: string
                  format: uuid
                  description: If provided, location is also persisted to the audit trail
      responses:
        '200':
          description: Location recorded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          recorded: { type: boolean }
                          id: { type: string, format: uuid, description: Postgres record ID (only when deliveryId provided) }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403':
          description: Driver role required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          description: Rate limited (max 1 request per 2 seconds per driver)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  # ── Delivery Acceptance (Driver Matching) ─────────────────────────────────
...
  /admin/drivers:
    get:
      tags: [Admin — Drivers]
      summary: List drivers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: search
          in: query
          schema: { type: string }
        - name: vehicleType
          in: query
          schema: { type: string, enum: [motorcycle, car, van, truck] }
        - name: verified
          in: query
          schema: { type: string, enum: ['true', 'false'] }
        - name: affiliation
          in: query
          schema: { type: string, enum: [independent, carrier] }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /admin/drivers/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Admin — Drivers]
      summary: Get driver details
      responses:
        '200': { $ref: '#/components/responses/Success' }

  # ── Admin — Deliveries ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Clerk JWT
      description: |
        Clerk-issued JWT. Obtain via @clerk/expo (mobile) or @clerk/react-router (web).
        Include as: Authorization: Bearer <token>
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Success:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    Envelope:
      type: object
      description: Standard success envelope
      properties:
        data: {}
        error: { nullable: true, example: null }
        meta:
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/PaginationMeta'
    ErrorEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { nullable: true, example: null }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
        meta: { nullable: true, example: null }
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

    # ── Shared enums ─────────────────────────────────────────────────────────
...